requirements.txt
absl-py==0.13.0
amqp==5.0.6
astunparse==1.6.3
billiard==3.6.4.0
cached-property==1.5.2
cachetools==4.2.2
celery==5.1.1
certifi==2021.5.30
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
cycler==0.10.0
Flask==2.0.1
flatbuffers==1.12
gast==0.3.3
google-auth==1.32.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
imageai==2.1.6
importlib-metadata==4.5.0
itsdangerous==2.0.1
Jinja2==3.0.1
Keras==2.4.3
Keras-Preprocessing==1.1.2
keras-resnet==0.2.0
kiwisolver==1.3.1
kombu==5.1.0
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.3.2
numpy==1.19.3
oauthlib==3.1.1
opencv-python==4.5.1.48
opt-einsum==3.3.0
Pillow==7.0.0
prompt-toolkit==3.0.19
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.4.1
six==1.15.0
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.0
tensorflow-estimator==2.4.0
termcolor==1.1.0
typing-extensions==3.7.4.3
urllib3==1.26.5
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.0.1
wrapt==1.12.1
zipp==3.4.1
